message box restructure
